add cleanup streams hook
